split hybrid.cpp
